fix: [win]Enable windows building after UI changed
After the UI change which support mobile for UOS, it blocks windows
building, remove phone classes from cooperation-transfer.

Log: Enable windows building after UI changed.
