+++ src/lib/cooperation/core/cooperationcoreplugin.h
@@ -27,7 +27,9 @@ public:
 private:
     void initialize();
     bool isMinilize();
+#ifdef ENABLE_PHONE
     void initMobileModule();
+#endif
     QSharedPointer<MainWindow> dMain { nullptr };
     bool onlyTransfer { false };
 };
