+++ src/lib/cooperation/core/gui/mainwindow.h
@@ -31,8 +31,9 @@ public:
     void closeEvent(QCloseEvent *event) override;
 
     void minimizedAPP();
-
+#ifdef ENABLE_PHONE
     void addMobileOperation(const QVariantMap &map);
+#endif
 
 signals:
     void refreshDevices();
@@ -49,10 +50,12 @@ public Q_SLOTS:
     void onRegistOperations(const QVariantMap &map);
     void onSwitchMode(CooperationMode mode);
 
+#ifdef ENABLE_PHONE
     //mobile
     void addMobileDevice(const DeviceInfoPointer info);
     void disconnectMobile();
     void onSetQRCode(const QString &code);
+#endif
 
 protected:
     void showCloseDialog();
