+++ src/lib/cooperation/transfer/CMakeLists.txt
@@ -39,8 +39,6 @@ FILE(GLOB PLUGIN_FILES
     "${CORE_UI_DIR}/gui/dialogs/*.cpp"
     "${CORE_UI_DIR}/gui/widgets/*.h"
     "${CORE_UI_DIR}/gui/widgets/*.cpp"
-    "${CORE_UI_DIR}/gui/phone/*.h"
-    "${CORE_UI_DIR}/gui/phone/*.cpp"
     "${CORE_UI_DIR}/gui/utils/*"
     "${CORE_UI_DIR}/discover/deviceinfo.h"
     "${CORE_UI_DIR}/discover/deviceinfo.cpp"
@@ -78,11 +76,6 @@ endif()
 
 find_package(QT NAMES Qt6 Qt5 REQUIRED COMPONENTS Core Gui Network DBus Concurrent)
 find_package(Qt${QT_VERSION_MAJOR} REQUIRED COMPONENTS Core Gui Widgets Network DBus Concurrent)
-
-find_library(QRENCODE_LIBRARY qrencode)
-if(NOT QRENCODE_LIBRARY)
-    message(FATAL_ERROR "libqrencode not found. Please install it.")
-endif()
 
 # 根据Qt版本选择合适的命令
 if (QT_VERSION VERSION_GREATER_EQUAL "6.0")
@@ -132,7 +125,6 @@ target_link_libraries(${PROJECT_NAME}
     Qt${QT_VERSION_MAJOR}::Network
     Qt${QT_VERSION_MAJOR}::Concurrent
     CuteIPC
-    qrencode
 )
 
 target_link_libraries(${PROJECT_NAME}
